+++ PiOSGravitybridge.js
@@ -1,0 +1,783 @@
+ 1) Relaying to Cronos (or other networks) via @cosmjs/stargate.
+ *    2) Gravity Bridge to Cosmos.
+ *    3) Pi mining simulation logic.
+ *
+ *  Author:     Josef Kurk Edwards
+ *  Maintainer: interchain.io and Stellar
+ *  License:    MIT
+ ***************************************************/
+
+require("dotenv").config();
+const axios = require("axios");
+const { ProposalClient } = require("pi-governance-sdk");    // Hypothetical Pi governance client
+const { StargateClient } = require("@cosmjs/stargate");      // For relaying to IBC-compatible chains
+const { DirectSecp256k1HdWallet } = require("@cosmjs/proto-signing");
+const StellarSdk = require("stellar-sdk");                   // For Pi wallet generation/validation
+const { GravityBridgeClient } = require("gravity-bridge-sdk");  // Hypothetical Gravity Bridge SDK
+
+/* -------------------------- ENV & SETUP -------------------------- */
+
+/**
+ * Check that essential ENV variables are set. 
+ * Add more checks if needed.
+ */
+function verifyEnv() {
+  const requiredVars = ["PIP_API_KEY", "PI_API_KEY", "MNEMONIC", "SENDER_ADDRESS", "ETHEREUM_RPC_URL", "COSMOS_RPC_URL", "PRIVATE_KEY"];
+  requiredVars.forEach((key) => {
+    if (!process.env[key]) {
+      console.warn(`⚠️ Missing environment variable: ${key}`);
+    }
+  });
+}
+verifyEnv();
+
+/* ---------------------- 1. PI WALLET TETHERING -------------------- */
+
+/**
+ * Generates a Pi-style wallet keypair (Stellar-based).
+ * Public keys start with 'G', secret keys start with 'S'.
+ */
+function generatePiWallet() {
+  const keypair = StellarSdk.Keypair.random();
+  return {
+    publicKey: keypair.publicKey(), // e.g. 'G...'
+    secretKey: keypair.secret(),    // e.g. 'S...'
+  };
+}
+
+/**
+ * Validates a Pi/stellar address (public key).
+ * Returns true if it's a valid Ed25519 public key in Stellar format.
+ */
+function validatePiAddress(address) {
+  return StellarSdk.StrKey.isValidEd25519PublicKey(address);
+}
+
+/**
+ * Fetches the homeDomain set for a Pi/stellar address 
+ * by querying a Horizon server. 
+ * For actual Pi, you'd change the server endpoint to Pi's Horizon.
+ */
+async function getHomeDomain(address) {
+  if (!validatePiAddress(address)) {
+    throw new Error(`Invalid Pi/stellar address: ${address}`);
+  }
+  const server = new StellarSdk.Server("https://horizon.stellar.org");
+  const account = await server.loadAccount(address);
+  return account.home_domain || "No home domain set";
+}
+
+/* -------------------- 2. GOVERNANCE PROPOSAL ---------------------- */
+
+// Example proposal details (replace with your real data)
+const proposal = {
+  title: "Integrate Pi, BTC, ETH Across Multi-Chains via BNB Bridge",
+  description: `
+    Proposal to integrate Pi, BTC, and ETH with Cronos, Ethereum,
+    Bitcoin, Linea, BSC, and Base ecosystems using Pi Network 
+    tokenomics. Pi will serve as the universal liquidity token 
+    while adhering to the mining, lockup, and utility incentives 
+    outlined in the Pi Whitepaper.
+  `,
+  proposer: "Josef Kurk Edwards",
+  fundingAmount: "1000000",
+  governanceType: "Integration Proposal",
+};
+
+/**
+ * Submits a governance proposal using the pi-governance-sdk.
+ */
+async function submitGovernanceProposal() {
+  try {
+    const client = new ProposalClient({
+      // Some references used process.env.PIP_API_KEY or PI_API_KEY
+      // Adjust as appropriate
+      apiKey: process.env.PIP_API_KEY || process.env.PI_API_KEY,
+      governanceEndpoint: "https://governance.minepi.com",
+    });
+
+    console.log("📢 Submitting governance proposal...");
+    const response = await client.submitProposal(proposal);
+
+    if (response.status === "success") {
+      console.log(`✅ Proposal submitted successfully! Proposal ID: ${response.proposalId}`);
+      return response.proposalId;
+    } else {
+      console.error(`❌ Proposal submission failed: ${response.message}`);
+    }
+  } catch (error) {
+    console.error("🚨 Error submitting proposal:", error.message);
+  }
+}
+
+/* ---------------------- 3. BRIDGE TO BNB  ------------------------- */
+
+const BINANCE_BRIDGE_API = "https://api.binance.org/bridge/v2";
+
+/**
+ * Bridges a token (BTC, ETH, PI, etc.) from its native chain to BNB Chain
+ * using the Binance Bridge API.
+ */
+async function bridgeTokenToBNB(tokenSymbol, amount, recipientAddress) {
+  try {
+    console.log(`🔄 Bridging ${tokenSymbol} -> BNB | Amount: ${amount}`);
+    const response = await axios.post(`${BINANCE_BRIDGE_API}/transfer`, {
+      asset: tokenSymbol,
+      amount: amount,
+      recipientAddress: recipientAddress,
+      network: "BNB",
+    });
+
+    if (response.data?.success) {
+      const txId = response.data.transactionId;
+      console.log(`✅ ${tokenSymbol} bridged successfully to BNB! TxID: ${txId}`);
+      return txId;
+    } else {
+      console.error(`❌ BNB Bridge failed: ${response.data?.message || "Unknown error"}`);
+    }
+  } catch (error) {
+    console.error("🚨 Error during bridging to BNB:", error.message);
+  }
+}
+
+/* -------------------- 4. RELAY TO CRONOS (or IBC) ----------------- */
+
+/**
+ * Example of relaying tokens from BNB to Cronos (or an IBC-compatible chain).
+ * 
+ * If Cronos is IBC-compatible, we'd craft an IBC MsgTransfer. 
+ * Otherwise, adapt to the bridging approach Cronos actually uses.
+ */
+async function relayToCronos(tokenSymbol, amount, recipientAddress) {
+  try {
+    console.log(`🚀 Relaying ${tokenSymbol} -> Cronos | Amount: ${amount}`);
+
+    // Example IBC-like message
+    const relayMsg = {
+      sourcePort: "transfer",
+      sourceChannel: "channel-0",  // Replace with Cronos’s actual channel ID to BNB chain
+      token: { denom: tokenSymbol, amount: String(amount) },
+      sender: process.env.SENDER_ADDRESS,
+      receiver: recipientAddress,
+    };
+
+    // Log for clarity
+    console.log("Relay message (Cronos):", relayMsg);
+
+    // Connect to Cronos with Stargate
+    // Replace 'https://rpc.cronos.org' with the real endpoint
+    const client = await StargateClient.connect("https://rpc.cronos.org");
+    const wallet = await DirectSecp256k1HdWallet.fromMnemonic(process.env.MNEMONIC);
+    const [account] = await wallet.getAccounts();
+
+    // Example fee; adjust for Cronos
+    const fee = {
+      amount: [{ denom: "basecro", amount: "5000" }],
+      gas: "200000",
+    };
+
+    const result = await client.signAndBroadcast(account.address, [relayMsg], fee);
+    if (result.code === 0) {
+      console.log(`✅ ${tokenSymbol} relayed to Cronos. Tx Hash: ${result.transactionHash}`);
+    } else {
+      console.error(`❌ Relay to Cronos failed: ${result.rawLog}`);
+    }
+  } catch (error) {
+    console.error("🚨 Error during relay to Cronos:", error.message);
+  }
+}
+
+/* ------------------- 5. GRAVITY BRIDGE TO COSMOS ------------------ */
+
+/**
+ * Bridges a token from Ethereum to Cosmos using Gravity Bridge.
+ * 
+ * @param {string} tokenContract - The Ethereum contract address of the token.
+ * @param {string} amount - The amount of tokens to bridge.
+ * @param {string} recipientAddress - The recipient address on the Cosmos chain.
+ * @returns {string} The transaction hash of the Gravity Bridge transfer.
+ */
+async function bridgeTokenViaGravity(tokenContract, amount, recipientAddress) {
+  try {
+    console.log(`🔄 Bridging ${tokenContract} -> Cosmos via Gravity | Amount: ${amount}`);
+
+    // Initialize Gravity Bridge client
+    const gravityClient = new GravityBridgeClient({
+      ethereumRpc: process.env.ETHEREUM_RPC_URL, // Ethereum RPC URL
+      cosmosRpc: process.env.COSMOS_RPC_URL,     // Cosmos RPC URL
+      privateKey: process.env.PRIVATE_KEY,       // Your Ethereum private key for signing transactions
+    });
+
+    // Perform the bridging operation
+    const txHash = await gravityClient.bridgeToCosmos({
+      tokenContract: tokenContract,
+      amount: amount,
+      recipient: recipientAddress,
+    });
+
+    console.log(`✅ Token bridged successfully via Gravity! Tx Hash: ${txHash}`);
+    return txHash;
+  } catch (error) {
+    console.error("🚨 Error during Gravity Bridge transfer:", error.message);
+  }
+}
+
+/* ------------------- 6. PI MINING SIMULATION -----------------------
+
+require("dotenv").config();
+const axios = require("axios");
+const { ProposalClient } = require("pi-governance-sdk");    // Hypothetical Pi governance client
+const { StargateClient } = require("@cosmjs/stargate");      // For relaying to IBC-compatible chains
+const { DirectSecp256k1HdWallet } = require("@cosmjs/proto-signing");
+const StellarSdk = require("stellar-sdk");                   // For Pi wallet generation/validation
+const { GravityBridgeClient } = require("gravity-bridge-sdk");  // Hypothetical Gravity Bridge SDK
+
+/* -------------------------- ENV & SETUP -------------------------- */
+
+/**
+ * Check that essential ENV variables are set. 
+ * Add more checks if needed.
+ */
+function verifyEnv() {
+  const requiredVars = ["PIP_API_KEY", "PI_API_KEY", "MNEMONIC", "SENDER_ADDRESS", "ETHEREUM_RPC_URL", "COSMOS_RPC_URL", "PRIVATE_KEY"];
+  requiredVars.forEach((key) => {
+    if (!process.env[key]) {
+      console.warn(`⚠️ Missing environment variable: ${key}`);
+    }
+  });
+}
+verifyEnv();
+
+/* ---------------------- 1. PI WALLET TETHERING -------------------- */
+
+/**
+ * Generates a Pi-style wallet keypair (Stellar-based).
+ * Public keys start with 'G', secret keys start with 'S'.
+ */
+function generatePiWallet() {
+  const keypair = StellarSdk.Keypair.random();
+  return {
+    publicKey: keypair.publicKey(), // e.g. 'G...'
+    secretKey: keypair.secret(),    // e.g. 'S...'
+  };
+}
+
+/**
+ * Validates a Pi/stellar address (public key).
+ * Returns true if it's a valid Ed25519 public key in Stellar format.
+ */
+function validatePiAddress(address) {
+  return StellarSdk.StrKey.isValidEd25519PublicKey(address);
+}
+
+/**
+ * Fetches the homeDomain set for a Pi/stellar address 
+ * by querying a Horizon server. 
+ * For actual Pi, you'd change the server endpoint to Pi's Horizon.
+ */
+async function getHomeDomain(address) {
+  if (!validatePiAddress(address)) {
+    throw new Error(`Invalid Pi/stellar address: ${address}`);
+  }
+  const server = new StellarSdk.Server("https://horizon.stellar.org");
+  const account = await server.loadAccount(address);
+  return account.home_domain || "No home domain set";
+}
+
+/* -------------------- 2. GOVERNANCE PROPOSAL ---------------------- */
+
+// Example proposal details (replace with your real data)
+const proposal = {
+  title: "Integrate Pi, BTC, ETH Across Multi-Chains via BNB Bridge",
+  description: `
+    Proposal to integrate Pi, BTC, and ETH with Cronos, Ethereum,
+    Bitcoin, Linea, BSC, and Base ecosystems using Pi Network 
+    tokenomics. Pi will serve as the universal liquidity token 
+    while adhering to the mining, lockup, and utility incentives 
+    outlined in the Pi Whitepaper.
+  `,
+  proposer: "Josef Kurk Edwards",
+  fundingAmount: "1000000",
+  governanceType: "Integration Proposal",
+};
+
+/**
+ * Submits a governance proposal using the pi-governance-sdk.
+ */
+async function submitGovernanceProposal() {
+  try {
+    const client = new ProposalClient({
+      // Some references used process.env.PIP_API_KEY or PI_API_KEY
+      // Adjust as appropriate
+      apiKey: process.env.PIP_API_KEY || process.env.PI_API_KEY,
+      governanceEndpoint: "https://governance.minepi.com",
+    });
+
+    console.log("📢 Submitting governance proposal...");
+    const response = await client.submitProposal(proposal);
+
+    if (response.status === "success") {
+      console.log(`✅ Proposal submitted successfully! Proposal ID: ${response.proposalId}`);
+      return response.proposalId;
+    } else {
+      console.error(`❌ Proposal submission failed: ${response.message}`);
+    }
+  } catch (error) {
+    console.error("🚨 Error submitting proposal:", error.message);
+  }
+}
+
+/* ---------------------- 3. BRIDGE TO BNB  ------------------------- */
+
+const BINANCE_BRIDGE_API = "https://api.binance.org/bridge/v2";
+
+/**
+ * Bridges a token (BTC, ETH, PI, etc.) from its native chain to BNB Chain
+ * using the Binance Bridge API.
+ */
+async function bridgeTokenToBNB(tokenSymbol, amount, recipientAddress) {
+  try {
+    console.log(`🔄 Bridging ${tokenSymbol} -> BNB | Amount: ${amount}`);
+    const response = await axios.post(`${BINANCE_BRIDGE_API}/transfer`, {
+      asset: tokenSymbol,
+      amount: amount,
+      recipientAddress: recipientAddress,
+      network: "BNB",
+    });
+
+    if (response.data?.success) {
+      const txId = response.data.transactionId;
+      console.log(`✅ ${tokenSymbol} bridged successfully to BNB! TxID: ${txId}`);
+      return txId;
+    } else {
+      console.error(`❌ BNB Bridge failed: ${response.data?.message || "Unknown error"}`);
+    }
+  } catch (error) {
+    console.error("🚨 Error during bridging to BNB:", error.message);
+  }
+}
+
+/* -------------------- 4. RELAY TO CRONOS (or IBC) ----------------- */
+
+/**
+ * Example of relaying tokens from BNB to Cronos (or an IBC-compatible chain).
+ * 
+ * If Cronos is IBC-compatible, we'd craft an IBC MsgTransfer. 
+ * Otherwise, adapt to the bridging approach Cronos actually uses.
+ */
+async function relayToCronos(tokenSymbol, amount, recipientAddress) {
+  try {
+    console.log(`🚀 Relaying ${tokenSymbol} -> Cronos | Amount: ${amount}`);
+
+    // Example IBC-like message
+    const relayMsg = {
+      sourcePort: "transfer",
+      sourceChannel: "channel-0",  // Replace with Cronos’s actual channel ID to BNB chain
+      token: { denom: tokenSymbol, amount: String(amount) },
+      sender: process.env.SENDER_ADDRESS,
+      receiver: recipientAddress,
+    };
+
+    // Log for clarity
+    console.log("Relay message (Cronos):", relayMsg);
+
+    // Connect to Cronos with Stargate
+    // Replace 'https://rpc.cronos.org' with the real endpoint
+    const client = await StargateClient.connect("https://rpc.cronos.org");
+    const wallet = await DirectSecp256k1HdWallet.fromMnemonic(process.env.MNEMONIC);
+    const [account] = await wallet.getAccounts();
+
+    // Example fee; adjust for Cronos
+    const fee = {
+      amount: [{ denom: "basecro", amount: "5000" }],
+      gas: "200000",
+    };
+
+    const result = await client.signAndBroadcast(account.address, [relayMsg], fee);
+    if (result.code === 0) {
+      console.log(`✅ ${tokenSymbol} relayed to Cronos. Tx Hash: ${result.transactionHash}`);
+    } else {
+      console.error(`❌ Relay to Cronos failed: ${result.rawLog}`);
+    }
+  } catch (error) {
+    console.error("🚨 Error during relay to Cronos:", error.message);
+  }
+}
+
+/* ------------------- 5. GRAVITY BRIDGE TO COSMOS ------------------ */
+
+/**
+ * Bridges a token from Ethereum to Cosmos using Gravity Bridge.
+ * 
+ * @param {string} tokenContract - The Ethereum contract address of the token.
+ * @param {string} amount - The amount of tokens to bridge.
+ * @param {string} recipientAddress - The recipient address on the Cosmos chain.
+ * @returns {string} The transaction hash of the Gravity Bridge transfer.
+ */
+async function bridgeTokenViaGravity(tokenContract, amount, recipientAddress) {
+  try {
+    console.log(`🔄 Bridging ${tokenContract} -> Cosmos via Gravity | Amount: ${amount}`);
+
+    // Initialize Gravity Bridge client
+    const gravityClient = new GravityBridgeClient({
+      ethereumRpc: process.env.ETHEREUM_RPC_URL, // Ethereum RPC URL
+      cosmosRpc: process.env.COSMOS_RPC_URL,     // Cosmos RPC URL
+      privateKey: process.env.PRIVATE_KEY,       // Your Ethereum private key for signing transactions
+    });
+
+    // Perform the bridging operation
+    const txHash = await gravityClient.bridgeToCosmos({
+      tokenContract: tokenContract,
+      amount: amount,
+      recipient: recipientAddress,
+    });
+
+    console.log(`✅ Token bridged successfully via Gravity! Tx Hash: ${txHash}`);
+    return txHash;
+  } catch (error) {
+    console.error("🚨 Error during Gravity Bridge transfer:", error.message);
+  }
+}
+
+/* ------------------- 6. PI MINING SIMULATION ----------------------- */
+
+const PI_MINING_BASE_RATE = 3.1415926;
+const LOCKUP_MULTIPLIERS = { 2: 0.1, 6
+
+/***************************************************
+ *  Integrated Pi-BTC-ETH Multi-Chain Script
+ *
+ *  Demonstrates:
+ *    1) Pi wallet tethering (Stellar-based "G" addresses).
+ *    2) Governance proposal submission (pi-governance-sdk).
+ *    3) Bridging to BNB (Binance Bridge).
+ *    4) Relaying to Cronos (or other networks) via @cosmjs/stargate.
+ *    5) Gravity Bridge to Cosmos.
+ *    6) Pi mining simulation logic.
+ *
+ *  Author:     Josef Kurk Edwards
+ *  Maintainer: You
+ *  License:    MIT
+ ***************************************************/
+
+require("dotenv").config();
+const axios = require("axios");
+const { ProposalClient } = require("pi-governance-sdk");    // Hypothetical Pi governance client
+const { StargateClient } = require("@cosmjs/stargate");      // For relaying to IBC-compatible chains
+const { DirectSecp256k1HdWallet } = require("@cosmjs/proto-signing");
+const StellarSdk = require("stellar-sdk");                   // For Pi wallet generation/validation
+const { GravityBridgeClient } = require("gravity-bridge-sdk");  // Hypothetical Gravity Bridge SDK
+
+/* -------------------------- ENV & SETUP -------------------------- */
+
+/**
+ * Check that essential ENV variables are set. 
+ * Add more checks if needed.
+ */
+function verifyEnv() {
+  const requiredVars = ["PIP_API_KEY", "PI_API_KEY", "MNEMONIC", "SENDER_ADDRESS", "ETHEREUM_RPC_URL", "COSMOS_RPC_URL", "PRIVATE_KEY"];
+  requiredVars.forEach((key) => {
+    if (!process.env[key]) {
+      console.warn(`⚠️ Missing environment variable: ${key}`);
+    }
+  });
+}
+verifyEnv();
+
+/* ---------------------- 1. PI WALLET TETHERING -------------------- */
+
+/**
+ * Generates a Pi-style wallet keypair (Stellar-based).
+ * Public keys start with 'G', secret keys start with 'S'.
+ */
+function generatePiWallet() {
+  const keypair = StellarSdk.Keypair.random();
+  return {
+    publicKey: keypair.publicKey(), // e.g. 'G...'
+    secretKey: keypair.secret(),    // e.g. 'S...'
+  };
+}
+
+/**
+ * Validates a Pi/stellar address (public key).
+ * Returns true if it's a valid Ed25519 public key in Stellar format.
+ */
+function validatePiAddress(address) {
+  return StellarSdk.StrKey.isValidEd25519PublicKey(address);
+}
+
+/**
+ * Fetches the homeDomain set for a Pi/stellar address 
+ * by querying a Horizon server. 
+ * For actual Pi, you'd change the server endpoint to Pi's Horizon.
+ */
+async function getHomeDomain(address) {
+  if (!validatePiAddress(address)) {
+    throw new Error(`Invalid Pi/stellar address: ${address}`);
+  }
+  const server = new StellarSdk.Server("https://horizon.stellar.org");
+  const account = await server.loadAccount(address);
+  return account.home_domain || "No home domain set";
+}
+
+/* -------------------- 2. GOVERNANCE PROPOSAL ---------------------- */
+
+// Example proposal details (replace with your real data)
+const proposal = {
+  title: "Integrate Pi, BTC, ETH Across Multi-Chains via BNB Bridge",
+  description: `
+    Proposal to integrate Pi, BTC, and ETH with Cronos, Ethereum,
+    Bitcoin, Linea, BSC, and Base ecosystems using Pi Network 
+    tokenomics. Pi will serve as the universal liquidity token 
+    while adhering to the mining, lockup, and utility incentives 
+    outlined in the Pi Whitepaper.
+  `,
+  proposer: "Josef Kurk Edwards",
+  fundingAmount: "1000000",
+  governanceType: "Integration Proposal",
+};
+
+/**
+ * Submits a governance proposal using the pi-governance-sdk.
+ */
+async function submitGovernanceProposal() {
+  try {
+    const client = new ProposalClient({
+      // Some references used process.env.PIP_API_KEY or PI_API_KEY
+      // Adjust as appropriate
+      apiKey: process.env.PIP_API_KEY || process.env.PI_API_KEY,
+      governanceEndpoint: "https://governance.minepi.com",
+    });
+
+    console.log("📢 Submitting governance proposal...");
+    const response = await client.submitProposal(proposal);
+
+    if (response.status === "success") {
+      console.log(`✅ Proposal submitted successfully! Proposal ID: ${response.proposalId}`);
+      return response.proposalId;
+    } else {
+      console.error(`❌ Proposal submission failed: ${response.message}`);
+    }
+  } catch (error) {
+    console.error("🚨 Error submitting proposal:", error.message);
+  }
+}
+
+/* ---------------------- 3. BRIDGE TO BNB  ------------------------- */
+
+const BINANCE_BRIDGE_API = "https://api.binance.org/bridge/v2";
+
+/**
+ * Bridges a token (BTC, ETH, PI, etc.) from its native chain to BNB Chain
+ * using the Binance Bridge API.
+ */
+async function bridgeTokenToBNB(tokenSymbol, amount, recipientAddress) {
+  try {
+    console.log(`🔄 Bridging ${tokenSymbol} -> BNB | Amount: ${amount}`);
+    const response = await axios.post(`${BINANCE_BRIDGE_API}/transfer`, {
+      asset: tokenSymbol,
+      amount: amount,
+      recipientAddress: recipientAddress,
+      network: "BNB",
+    });
+
+    if (response.data?.success) {
+      const txId = response.data.transactionId;
+      console.log(`✅ ${tokenSymbol} bridged successfully to BNB! TxID: ${txId}`);
+      return txId;
+    } else {
+      console.error(`❌ BNB Bridge failed: ${response.data?.message || "Unknown error"}`);
+    }
+  } catch (error) {
+    console.error("🚨 Error during bridging to BNB:", error.message);
+  }
+}
+
+/* -------------------- 4. RELAY TO CRONOS (or IBC) ----------------- */
+
+/**
+ * Example of relaying tokens from BNB to Cronos (or an IBC-compatible chain).
+ * 
+ * If Cronos is IBC-compatible, we'd craft an IBC MsgTransfer. 
+ * Otherwise, adapt to the bridging approach Cronos actually uses.
+ */
+async function relayToCronos(tokenSymbol, amount, recipientAddress) {
+  try {
+    console.log(`🚀 Relaying ${tokenSymbol} -> Cronos | Amount: ${amount}`);
+
+    // Example IBC-like message
+    const relayMsg = {
+      sourcePort: "transfer",
+      sourceChannel: "channel-0",  // Replace with Cronos’s actual channel ID to BNB chain
+      token: { denom: tokenSymbol, amount: String(amount) },
+      sender: process.env.SENDER_ADDRESS,
+      receiver: recipientAddress,
+    };
+
+    // Log for clarity
+    console.log("Relay message (Cronos):", relayMsg);
+
+    // Connect to Cronos with Stargate
+    // Replace 'https://rpc.cronos.org' with the real endpoint
+    const client = await StargateClient.connect("https://rpc.cronos.org");
+    const wallet = await DirectSecp256k1HdWallet.fromMnemonic(process.env.MNEMONIC);
+    const [account] = await wallet.getAccounts();
+
+    // Example fee; adjust for Cronos
+    const fee = {
+      amount: [{ denom: "basecro", amount: "5000" }],
+      gas: "200000",
+    };
+
+    const result = await client.signAndBroadcast(account.address, [relayMsg], fee);
+    if (result.code === 0) {
+      console.log(`✅ ${tokenSymbol} relayed to Cronos. Tx Hash: ${result.transactionHash}`);
+    } else {
+      console.error(`❌ Relay to Cronos failed: ${result.rawLog}`);
+    }
+  } catch (error) {
+    console.error("🚨 Error during relay to Cronos:", error.message);
+  }
+}
+
+/* ------------------- 5. GRAVITY BRIDGE TO COSMOS ------------------ */
+
+/**
+ * Bridges a token from Ethereum to Cosmos using Gravity Bridge.
+ * 
+ * @param {string} tokenContract - The Ethereum contract address of the token.
+ * @param {string} amount - The amount of tokens to bridge.
+ * @param {string} recipientAddress - The recipient address on the Cosmos chain.
+ * @returns {string} The transaction hash of the Gravity Bridge transfer.
+ */
+async function bridgeTokenViaGravity(tokenContract, amount, recipientAddress) {
+  try {
+    console.log(`🔄 Bridging ${tokenContract} -> Cosmos via Gravity | Amount: ${amount}`);
+
+    // Initialize Gravity Bridge client
+    const gravityClient = new GravityBridgeClient({
+      ethereumRpc: process.env.ETHEREUM_RPC_URL, // Ethereum RPC URL
+      cosmosRpc: process.env.COSMOS_RPC_URL,     // Cosmos RPC URL
+      privateKey: process.env.PRIVATE_KEY,       // Your Ethereum private key for signing transactions
+    });
+
+    // Perform the bridging operation
+    const txHash = await gravityClient.bridgeToCosmos({
+      tokenContract: tokenContract,
+      amount: amount,
+      recipient: recipientAddress,
+    });
+
+    console.log(`✅ Token bridged successfully via Gravity! Tx Hash: ${txHash}`);
+    return txHash;
+  } catch (error) {
+    console.error("🚨 Error during Gravity Bridge transfer:", error.message);
+  }
+}
+
+/* ------------------- 6. PI MINING SIMULATION -----------------------
+const PI_MINING_BASE_RATE = 3.1415926;
+const LOCKUP_MULTIPLIERS = { 2: 0.1, 6: 0.5, 12: 1, 36: 2 };
+/**
+ * Calculate the final mining rate (M) for Pi.
+ * @param {number} B - Base mining rate.
+ * @param {number} S - Security circle bonus.
+ * @param {number} L - Lockup reward portion.
+ * @param {number} E - Ecosystem bonus.
+ * @param {number} N - Network bonus.
+ * @param {number} A - Active engagement bonus.
+ * @param {number} X - Node bonus.
+ * @returns {number} - Final mining rate.
+ */
+function calculateMiningRate(B, S, L, E, N, A, X) {
+  return B * (1 + S + L) * (1 + N + E + A + X);
+}
+
+/**
+ * Calculate the lockup reward portion (L).
+ * @param {number} lockupPercentage - Percentage of Pi locked up.
+ * @param {number} lockupDurationMonths - Duration of lockup in months.
+ * @param {number} miningHistory - Historical mining data.
+ * @returns {number} - Lockup reward multiplier.
+ */
+function calculateLockupReward(lockupPercentage, lockupDurationMonths, miningHistory) {
+  const Lp = lockupPercentage / 100;
+  const Lt = LOCKUP_MULTIPLIERS[lockupDurationMonths] || 0;
+  const Mh = miningHistory || 1; // Assuming a base multiplier from history
+  return Lp * Lt * Mh;
+}
+
+/**
+ * Simulate Pi mining based on various factors and bonuses.
+ * @param {number} baseRate - Base mining rate.
+ * @param {number} securityBonus - Security circle bonus.
+ * @param {number} lockupPercentage - Percentage of Pi locked up.
+ * @param {number} lockupDuration - Duration of lockup in months.
+ * @param {number} ecosystemBonus - Ecosystem bonus.
+ * @param {number} networkBonus - Network bonus.
+ * @param {number} engagementBonus - Active engagement bonus.
+ * @param {number} nodeBonus - Node bonus.
+ * @returns {number} - Calculated Pi mining rate.
+ */
+function simulatePiMining(baseRate, securityBonus, lockupPercentage, lockupDuration, ecosystemBonus, networkBonus, engagementBonus, nodeBonus) {
+  const lockupReward = calculateLockupReward(lockupPercentage, lockupDuration, 1); // Assuming a default mining history
+  return calculateMiningRate(baseRate, securityBonus, lockupReward, ecosystemBonus, networkBonus, engagementBonus, nodeBonus);
+}
+
+/* ------------------------ MAIN SCRIPT ---------------------------- */
+
+/**
+ * Main function to demonstrate the full flow.
+ */
+async function main() {
+  verifyEnv();
+
+  // 1. Pi Wallet Tethering
+  const piWallet = generatePiWallet();
+  console.log("Generated Pi Wallet:", piWallet);
+
+  // 2. Governance Proposal Submission
+  const proposalId = await submitGovernanceProposal();
+  if (!proposalId) return;
+
+  // 3. Bridge to BNB
+  const tokenSymbol = "BTC";
+  const amount = "0.01";
+  const recipientAddress = "bnb1.....";  // Replace with actual BNB recipient address
+  const txId = await bridgeTokenToBNB(tokenSymbol, amount, recipientAddress);
+  if (!txId) return;
+
+  // 4. Relay to Cronos
+  const cronosAddress = "cronos1.....";  // Replace with actual Cronos recipient address
+  await relayToCronos(tokenSymbol, amount, cronosAddress);
+
+  // 5. Gravity Bridge to Cosmos
+  const ethereumTokenContract = "0x.....";  // Replace with actual Ethereum token contract address
+  const cosmosRecipientAddress = "cosmos1.....";  // Replace with actual Cosmos recipient address
+  await bridgeTokenViaGravity(ethereumTokenContract, amount, cosmosRecipientAddress);
+
+  // 6. Pi Mining Simulation
+  const baseRate = PI_MINING_BASE_RATE;
+  const securityBonus = 0.1; // Example security bonus
+  const lockupPercentage = 50; // Example lockup percentage
+  const lockupDuration = 12; // Example lockup duration in months
+  const ecosystemBonus = 0.2; // Example ecosystem bonus
+  const networkBonus = 0.1; // Example network bonus
+  const engagementBonus = 0.05; // Example active engagement bonus
+  const nodeBonus = 0.15; // Example node bonus
+
+  const miningRate = simulatePiMining(baseRate, securityBonus, lockupPercentage, lockupDuration, ecosystemBonus, networkBonus, engagementBonus, nodeBonus);
+  console.log(`Calculated Pi Mining Rate: ${miningRate}`);
+}
+
+// Run the main script
+main().catch(console.error);
+  console.log("\n🎉 Multi-chain integration process completed!");
+}
+
+// Run it
+executeIntegration().catch((err) => {
+  console.error("🚨 Integration error:", err);
+  process.exit(1);
+});
+
