--- PiOSbridge.js
@@ -1,314 +1,0 @@
-/***************************************************
- *  Integrated Pi-BTC-ETH Multi-Chain Script
- *
- *  Demonstrates:
- *    1) Pi wallet tethering (Stellar-based "G" addresses).
- *    2) Governance proposal submission (pi-governance-sdk).
- *    3) Bridging to BNB (Binance Bridge).
- *    4) Relaying to Cronos (or other networks) via @cosmjs/stargate.
- *    5) Pi mining simulation logic.
- *
- *  Author:     Josef Kurk Edwards
- *  Maintainer: You
- *  License:    MIT
- ***************************************************/
-
-require("dotenv").config();
-const axios = require("axios");
-const { ProposalClient } = require("pi-governance-sdk");    // Hypothetical Pi governance client
-const { StargateClient } = require("@cosmjs/stargate");      // For relaying to IBC-compatible chains
-const { DirectSecp256k1HdWallet } = require("@cosmjs/proto-signing");
-const StellarSdk = require("stellar-sdk");                   // For Pi wallet generation/validation
-
-/* -------------------------- ENV & SETUP -------------------------- */
-
-/**
- * Check that essential ENV variables are set. 
- * Add more checks if needed.
- */
-function verifyEnv() {
-  const requiredVars = ["PIP_API_KEY", "PI_API_KEY", "MNEMONIC", "SENDER_ADDRESS"];
-  requiredVars.forEach((key) => {
-    if (!process.env[key]) {
-      console.warn(`⚠️ Missing environment variable: ${key}`);
-    }
-  });
-}
-verifyEnv();
-
-/* ---------------------- 1. PI WALLET TETHERING -------------------- */
-
-/**
- * Generates a Pi-style wallet keypair (Stellar-based).
- * Public keys start with 'G', secret keys start with 'S'.
- */
-function generatePiWallet() {
-  const keypair = StellarSdk.Keypair.random();
-  return {
-    publicKey: keypair.publicKey(), // e.g. 'G...'
-    secretKey: keypair.secret(),    // e.g. 'S...'
-  };
-}
-
-/**
- * Validates a Pi/stellar address (public key).
- * Returns true if it's a valid Ed25519 public key in Stellar format.
- */
-function validatePiAddress(address) {
-  return StellarSdk.StrKey.isValidEd25519PublicKey(address);
-}
-
-/**
- * Fetches the homeDomain set for a Pi/stellar address 
- * by querying a Horizon server. 
- * For actual Pi, you'd change the server endpoint to Pi's Horizon.
- */
-async function getHomeDomain(address) {
-  if (!validatePiAddress(address)) {
-    throw new Error(`Invalid Pi/stellar address: ${address}`);
-  }
-  const server = new StellarSdk.Server("https://horizon.stellar.org");
-  const account = await server.loadAccount(address);
-  return account.home_domain || "No home domain set";
-}
-
-/* -------------------- 2. GOVERNANCE PROPOSAL ---------------------- */
-
-// Example proposal details (replace with your real data)
-const proposal = {
-  title: "Integrate Pi, BTC, ETH Across Multi-Chains via BNB Bridge",
-  description: `
-    Proposal to integrate Pi, BTC, and ETH with Cronos, Ethereum,
-    Bitcoin, Linea, BSC, and Base ecosystems using Pi Network 
-    tokenomics. Pi will serve as the universal liquidity token 
-    while adhering to the mining, lockup, and utility incentives 
-    outlined in the Pi Whitepaper.
-  `,
-  proposer: "Josef Kurk Edwards",
-  fundingAmount: "1000000",
-  governanceType: "Integration Proposal",
-};
-
-/**
- * Submits a governance proposal using the pi-governance-sdk.
- */
-async function submitGovernanceProposal() {
-  try {
-    const client = new ProposalClient({
-      // Some references used process.env.PIP_API_KEY or PI_API_KEY
-      // Adjust as appropriate
-      apiKey: process.env.PIP_API_KEY || process.env.PI_API_KEY,
-      governanceEndpoint: "https://governance.minepi.com",
-    });
-
-    console.log("📢 Submitting governance proposal...");
-    const response = await client.submitProposal(proposal);
-
-    if (response.status === "success") {
-      console.log(`✅ Proposal submitted successfully! Proposal ID: ${response.proposalId}`);
-      return response.proposalId;
-    } else {
-      console.error(`❌ Proposal submission failed: ${response.message}`);
-    }
-  } catch (error) {
-    console.error("🚨 Error submitting proposal:", error.message);
-  }
-}
-
-/* ---------------------- 3. BRIDGE TO BNB  ------------------------- */
-
-const BINANCE_BRIDGE_API = "https://api.binance.org/bridge/v2";
-
-/**
- * Bridges a token (BTC, ETH, PI, etc.) from its native chain to BNB Chain
- * using the Binance Bridge API.
- */
-async function bridgeTokenToBNB(tokenSymbol, amount, recipientAddress) {
-  try {
-    console.log(`🔄 Bridging ${tokenSymbol} -> BNB | Amount: ${amount}`);
-    const response = await axios.post(`${BINANCE_BRIDGE_API}/transfer`, {
-      asset: tokenSymbol,
-      amount: amount,
-      recipientAddress: recipientAddress,
-      network: "BNB",
-    });
-
-    if (response.data?.success) {
-      const txId = response.data.transactionId;
-      console.log(`✅ ${tokenSymbol} bridged successfully to BNB! TxID: ${txId}`);
-      return txId;
-    } else {
-      console.error(`❌ BNB Bridge failed: ${response.data?.message || "Unknown error"}`);
-    }
-  } catch (error) {
-    console.error("🚨 Error during bridging to BNB:", error.message);
-  }
-}
-
-/* -------------------- 4. RELAY TO CRONOS (or IBC) ----------------- */
-
-/**
- * Example of relaying tokens from BNB to Cronos (or an IBC-compatible chain).
- * 
- * If Cronos is IBC-compatible, we'd craft an IBC MsgTransfer. 
- * Otherwise, adapt to the bridging approach Cronos actually uses.
- */
-async function relayToCronos(tokenSymbol, amount, recipientAddress) {
-  try {
-    console.log(`🚀 Relaying ${tokenSymbol} -> Cronos | Amount: ${amount}`);
-
-    // Example IBC-like message
-    const relayMsg = {
-      sourcePort: "transfer",
-      sourceChannel: "channel-0",  // Replace with Cronos’s actual channel ID to BNB chain
-      token: { denom: tokenSymbol, amount: String(amount) },
-      sender: process.env.SENDER_ADDRESS,
-      receiver: recipientAddress,
-    };
-
-    // Log for clarity
-    console.log("Relay message (Cronos):", relayMsg);
-
-    // Connect to Cronos with Stargate
-    // Replace 'https://rpc.cronos.org' with the real endpoint
-    const client = await StargateClient.connect("https://rpc.cronos.org");
-    const wallet = await DirectSecp256k1HdWallet.fromMnemonic(process.env.MNEMONIC);
-    const [account] = await wallet.getAccounts();
-
-    // Example fee; adjust for Cronos
-    const fee = {
-      amount: [{ denom: "basecro", amount: "5000" }],
-      gas: "200000",
-    };
-
-    const result = await client.signAndBroadcast(account.address, [relayMsg], fee);
-    if (result.code === 0) {
-      console.log(`✅ ${tokenSymbol} relayed to Cronos. Tx Hash: ${result.transactionHash}`);
-    } else {
-      console.error(`❌ Relay to Cronos failed: ${result.rawLog}`);
-    }
-  } catch (error) {
-    console.error("🚨 Error during relay to Cronos:", error.message);
-  }
-}
-
-/* ------------------- 5. PI MINING SIMULATION ----------------------- */
-
-const PI_MINING_BASE_RATE = 3.1415926;
-const LOCKUP_MULTIPLIERS = { 2: 0.1, 6: 0.5, 12: 1, 36: 2 };
-
-/**
- * Calculate the final mining rate (M) for Pi.
- */
-function calculateMiningRate(B, S, L, E, N, A, X) {
-  return B * (1 + S + L) * (1 + N + E + A + X);
-}
-
-/**
- * Calculate the lockup reward portion (L).
- */
-function calculateLockupReward(lockupPercentage, lockupDurationMonths, miningHistory) {
-  const Lp = lockupPercentage / 100;
-  const Lt = LOCKUP_MULTIPLIERS[lockupDurationMonths] || 0;
-  const N = Math.max(1, miningHistory);
-  const logN = Math.log(N);
-  return Lt * Lp * logN;
-}
-
-/**
- * Example function to simulate Pi mining for a single user.
- */
-function simulatePiMining(pioneer) {
-  const {
-    securityCircle,
-    lockupPercentage,
-    lockupDuration,
-    referralCount,
-    nodeUptime,
-    appUsage,
-    miningHistory,
-  } = pioneer;
-
-  // Example factors
-  const B = PI_MINING_BASE_RATE;
-  const S = 0.2 * Math.min(securityCircle, 5); 
-  const L = calculateLockupReward(lockupPercentage, lockupDuration, miningHistory);
-  const E = 0.25 * referralCount;
-  const N = nodeUptime > 0.5 ? 0.1 : 0;
-  const A = appUsage > 60 ? 0.05 : 0;
-  const X = 0; 
-
-  const M = calculateMiningRate(B, S, L, E, N, A, X);
-  console.log(`🛠️ Pioneer Mining Rate: ${M.toFixed(6)} Pi/h (simulated)`);
-  return M;
-}
-
-/* -------------------- MAIN EXECUTION FLOW -------------------------- */
-
-/**
- * Puts it all together:
- * 1) Submits governance proposal
- * 2) Bridges BTC, ETH, PI to BNB
- * 3) Relays them to Cronos
- * 4) Demonstrates Pi wallet tethering
- * 5) Simulates Pi mining
- */
-async function executeIntegration() {
-  console.log("🔧 Starting multi-chain integration process...");
-
-  // 1. Governance proposal
-  const proposalId = await submitGovernanceProposal();
-  if (!proposalId) {
-    console.warn("⚠️ No proposal ID returned. Continuing anyway...");
-  }
-
-  // 2. Bridge tokens to BNB
-  //    For demonstration, bridging small amounts
-  const bnbAddress = "bnb1recipientaddress..."; // Replace with real BNB address
-  await bridgeTokenToBNB("BTC", "0.01", bnbAddress);
-  await bridgeTokenToBNB("ETH", "1",   bnbAddress);
-  await bridgeTokenToBNB("PI",  "1000", bnbAddress);
-
-  // 3. Relay from BNB to Cronos
-  //    After bridging, your tokens are typically "BTCB", "ETHB", "PIB" on BNB side
-  const cronosAddress = "crc1recipientaddress..."; // Replace with real Cronos address
-  await relayToCronos("BTCB", "0.01", cronosAddress);
-  await relayToCronos("ETHB", "1",    cronosAddress);
-  await relayToCronos("PIB",  "1000", cronosAddress);
-
-  // 4. Pi wallet tethering demonstration
-  console.log("\n🔗 Generating & validating Pi wallets...");
-  const newPiWallet = generatePiWallet();
-  console.log("New Pi wallet:", newPiWallet);
-
-  const samplePiAddress = "GC7XIJBRK5C2E7IKND7HP3NMEKE435HNLLBG7STDSO2TXCJDW4B2KF25";
-  console.log(`Is ${samplePiAddress} valid?`, validatePiAddress(samplePiAddress));
-  try {
-    const domain = await getHomeDomain(samplePiAddress);
-    console.log(`Home domain for ${samplePiAddress}:`, domain);
-  } catch (err) {
-    console.warn("Home domain fetch error:", err.message);
-  }
-
-  // 5. Pi mining simulation
-  console.log("\n⛏️ Simulating Pi mining for a sample pioneer:");
-  const pioneer = {
-    securityCircle: 4,
-    lockupPercentage: 90,
-    lockupDuration: 12,
-    referralCount: 3,
-    nodeUptime: 0.7,
-    appUsage: 120,
-    miningHistory: 500,
-  };
-  simulatePiMining(pioneer);
-
-  console.log("\n🎉 Multi-chain integration process completed!");
-}
-
-// Run it
-executeIntegration().catch((err) => {
-  console.error("🚨 Integration error:", err);
-  process.exit(1);
-});
-
